+++ client/src/components/ui/Modal/BulkResultModal.tsx
